refactor: Remove deprecated media and news routes, and reorganize controller structure
