+++ backend/routes/newsRoute.js
